myLosson
This project is my first comprehensive program. Initially, users are required to register in order to access the program. After registration, they must log in. Once logged in, they will be able to access either developer or manager functions. The project is open for further development, and in future updates, a separate administrator password will be required to access management functions.